Changes the cache behaviour when the application is online The cache is only used when the application is offline and isn't anymore used to optimize infromation retrieval when the application is online The documentation of the database interface was also improved
