+++ app/src/androidTest/java/ch/epfl/sdp/kandle/LoginActivityTest.java
@@ -21,7 +21,7 @@ import androidx.test.rule.GrantPermissionRule;
 import java.util.HashMap;
 
 
-import ch.epfl.sdp.kandle.Storage.room.LocalDatabase;
+import ch.epfl.sdp.kandle.storage.room.LocalDatabase;
 import ch.epfl.sdp.kandle.activity.LoginActivity;
 
 import ch.epfl.sdp.kandle.activity.MainActivity;
