+++ app/src/androidTest/java/ch/epfl/sdp/kandle/LoginActivityTestAlreadyLoggedIn.java
@@ -11,7 +11,7 @@ import androidx.test.rule.ActivityTestRule;
 
 import java.util.HashMap;
 
-import ch.epfl.sdp.kandle.Storage.room.LocalDatabase;
+import ch.epfl.sdp.kandle.storage.room.LocalDatabase;
 import ch.epfl.sdp.kandle.activity.LoginActivity;
 import ch.epfl.sdp.kandle.dependencies.DependencyManager;
 import ch.epfl.sdp.kandle.dependencies.MockAuthentication;
