+++ app/src/androidTest/java/ch/epfl/sdp/kandle/MapViewFragmentTest.java
@@ -19,7 +19,7 @@ import org.junit.Test;
 import java.util.Date;
 import java.util.HashMap;
 
-import ch.epfl.sdp.kandle.Storage.room.LocalDatabase;
+import ch.epfl.sdp.kandle.storage.room.LocalDatabase;
 import ch.epfl.sdp.kandle.activity.MainActivity;
 import ch.epfl.sdp.kandle.dependencies.DependencyManager;
 import ch.epfl.sdp.kandle.dependencies.MockAuthentication;
