+++ app/src/androidTest/java/ch/epfl/sdp/kandle/PostActivityTest.java
@@ -18,7 +18,7 @@ import org.junit.runner.RunWith;
 
 import java.util.HashMap;
 
-import ch.epfl.sdp.kandle.Storage.room.LocalDatabase;
+import ch.epfl.sdp.kandle.storage.room.LocalDatabase;
 import ch.epfl.sdp.kandle.activity.PostActivity;
 import ch.epfl.sdp.kandle.dependencies.DependencyManager;
 import ch.epfl.sdp.kandle.dependencies.MockAuthentication;
