+++ app/src/androidTest/java/ch/epfl/sdp/kandle/SettingsFragmentTest.java
@@ -14,7 +14,7 @@ import androidx.test.rule.GrantPermissionRule;
 
 import java.util.HashMap;
 
-import ch.epfl.sdp.kandle.Storage.room.LocalDatabase;
+import ch.epfl.sdp.kandle.storage.room.LocalDatabase;
 import ch.epfl.sdp.kandle.activity.MainActivity;
 import ch.epfl.sdp.kandle.dependencies.DependencyManager;
 import ch.epfl.sdp.kandle.dependencies.MockAuthentication;
