+++ app/src/main/java/ch/epfl/sdp/kandle/User.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import ch.epfl.sdp.kandle.Storage.room.Converters;
+import ch.epfl.sdp.kandle.storage.room.Converters;
 
 @Entity(tableName = "Users")
 public class User implements Serializable {
