+++ app/src/main/java/ch/epfl/sdp/kandle/storage/caching/InternalStorageHandler.java
@@ -1,4 +1,4 @@
-package ch.epfl.sdp.kandle.Storage.caching;
+package ch.epfl.sdp.kandle.storage.caching;
 
 import android.content.Context;
 
@@ -19,10 +19,11 @@ import ch.epfl.sdp.kandle.dependencies.InternalStorage;
  * @Author Marc Egli
  * This class handles the internal storage accesses
  */
+@SuppressWarnings("ResultOfMethodCallIgnored")
 public class InternalStorageHandler implements InternalStorage {
 
     private final String USER_DATA_PATH = "userData";
-    private Context context;
+    private final Context context;
     private static final InternalStorageHandler INSTANCE = new InternalStorageHandler();
 
     public InternalStorageHandler() {
@@ -87,10 +88,6 @@ public class InternalStorageHandler implements InternalStorage {
     @Override
     public void saveUserAtLoginOrRegister(@NonNull User user) throws IllegalArgumentException {
 
-        if (user == null) {
-            throw new IllegalArgumentException();
-        }
-
         User storedUser = getCurrentUser();
         if (storedUser == null) {
             storeUser(user);
@@ -112,9 +109,6 @@ public class InternalStorageHandler implements InternalStorage {
 
     @Override
     public void updateUser(@NonNull User user) throws IllegalArgumentException {
-        if (user == null) {
-            throw new IllegalArgumentException();
-        }
         deleteUser();
         storeUser(user);
