+++ app/src/main/java/ch/epfl/sdp/kandle/storage/room/Converters.java
@@ -1,15 +1,13 @@
-package ch.epfl.sdp.kandle.Storage.room;
+package ch.epfl.sdp.kandle.storage.room;
 
 import androidx.annotation.Nullable;
 import androidx.room.TypeConverter;
-import androidx.room.TypeConverters;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Date;
 import java.util.List;
