+++ app/src/main/java/ch/epfl/sdp/kandle/storage/room/UserDao.java
@@ -1,8 +1,9 @@
-package ch.epfl.sdp.kandle.Storage.room;
+package ch.epfl.sdp.kandle.storage.room;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
 import androidx.room.Insert;
+import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 import androidx.room.Update;
 
@@ -15,8 +16,11 @@ public interface UserDao {
     @Query("SELECT * from Users")
     List<User> getUserList();
 
-    @Insert
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
     void insertUser(User user);
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    void insertAllUsers(List<User> users);
 
     @Update
     void updateUser(User user);
